+++ core/src/veeronten/hackathon/Utils.kt
@@ -31,6 +31,6 @@ fun createSlider(min: Int, max:Int, skin: Skin, onClick: (Int) -> Unit) = Slider
     }
 }
 
-fun Batch.drawByCoord(img: Texture, point: Point) = draw(img, TILE_SIZE.toFloat() * point.x, HEIGHT - TILE_SIZE.toFloat() * (point.y + 1), TILE_SIZE.toFloat(), TILE_SIZE.toFloat())
+fun Batch.drawByCoord(img: Texture, point: Point) = draw(img, TILE_SIZE * point.x, HEIGHT - TILE_SIZE * (point.y + 1), TILE_SIZE, TILE_SIZE)
 
 fun lerp(start: Float, end: Float, t: Float) = start + t * (end - start)
